# server/target/openapi.yaml
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "Control Finanzas API",
                 "description":  "API para cuentas y ahorro (iteraciÃ³n inicial)",
                 "version":  "0.1.0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Generated server url"
                    }
                ],
    "paths":  {
                  "/api/v1/budgets/categories/{categoryId}":  {
                                                                  "put":  {
                                                                              "tags":  [
                                                                                           "budget-controller"
                                                                                       ],
                                                                              "operationId":  "upsert",
                                                                              "parameters":  [

                                                                                             ],
                                                                              "requestBody":  {
                                                                                                  "content":  {
                                                                                                                  "application/json":  {
                                                                                                                                           "schema":  {
                                                                                                                                                          "$ref":  "#/components/schemas/UpsertBody"
                                                                                                                                                      }
                                                                                                                                       }
                                                                                                              },
                                                                                                  "required":  true
                                                                                              },
                                                                              "responses":  {
                                                                                                "200":  {
                                                                                                            "description":  "OK",
                                                                                                            "content":  {
                                                                                                                            "*/*":  {
                                                                                                                                        "schema":  {
                                                                                                                                                       "type":  "object"
                                                                                                                                                   }
                                                                                                                                    }
                                                                                                                        }
                                                                                                        }
                                                                                            }
                                                                          }
                                                              },
...
                  "/api/v1/categories":  {
                                             "get":  {
                                                         "tags":  [
                                                                      "category-controller"
                                                                  ],
                                                         "operationId":  "list_2",
                                                         "responses":  {
                                                                           "200":  {
                                                                                       "description":  "OK",
                                                                                       "content":  {
                                                                                                       "*/*":  {
                                                                                                                   "schema":  {
                                                                                                                                  "type":  "array",
                                                                                                                                  "items":  {
                                                                                                                                                "$ref":  "#/components/schemas/Category"
                                                                                                                                            }
                                                                                                                              }
                                                                                                               }
                                                                                                   }
                                                                                   }
                                                                       }
                                                     },
                                             "post":  {
                                                          "tags":  [
                                                                       "category-controller"
                                                                   ],
                                                          "operationId":  "create_2",
                                                          "requestBody":  {
                                                                              "content":  {
                                                                                              "application/json":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/Body"
                                                                                                                                  }
                                                                                                                   }
                                                                                          },
                                                                              "required":  true
                                                                          },
                                                          "responses":  {
                                                                            "200":  {
                                                                                        "description":  "OK",
                                                                                        "content":  {
                                                                                                        "*/*":  {
                                                                                                                    "schema":  {
                                                                                                                                   "type":  "object"
                                                                                                                               }
                                                                                                                }
                                                                                                    }
                                                                                    }
                                                                        }
                                                      }
                                         },
...
                  "/api/v1/categories/{id}":  {
                                                  "delete":  {
                                                                 "tags":  [
                                                                              "category-controller"
                                                                          ],
                                                                 "operationId":  "delete_2",
                                                                 "parameters":  [

                                                                                ],
                                                                 "responses":  {
                                                                                   "200":  {
                                                                                               "description":  "OK",
                                                                                               "content":  {
                                                                                                               "*/*":  {
                                                                                                                           "schema":  {
                                                                                                                                          "type":  "object"
                                                                                                                                      }
                                                                                                                       }
                                                                                                           }
                                                                                           }
                                                                               }
                                                             },
                                                  "patch":  {
                                                                "tags":  [
                                                                             "category-controller"
                                                                         ],
                                                                "operationId":  "rename",
                                                                "parameters":  [

                                                                               ],
                                                                "requestBody":  {
                                                                                    "content":  {
                                                                                                    "application/json":  {
                                                                                                                             "schema":  {
                                                                                                                                            "$ref":  "#/components/schemas/Body"
                                                                                                                                        }
                                                                                                                         }
                                                                                                },
                                                                                    "required":  true
                                                                                },
                                                                "responses":  {
                                                                                  "200":  {
                                                                                              "description":  "OK",
                                                                                              "content":  {
                                                                                                              "*/*":  {
                                                                                                                          "schema":  {
                                                                                                                                         "type":  "object"
                                                                                                                                     }
                                                                                                                      }
                                                                                                          }
                                                                                          }
                                                                              }
                                                            }
                                              },
...
              },
    "components":  {
                       "schemas":  {
                                       "UpsertBody":  {
                                                          "required":  [
                                                                           "amount",
                                                                           "categoryId",
                                                                           "monthKey"
                                                                       ],
                                                          "type":  "object",
                                                          "properties":  {
                                                                             "categoryId":  {
                                                                                                "type":  "integer",
                                                                                                "format":  "int64"
                                                                                            },
                                                                             "monthKey":  {
                                                                                              "type":  "string"
                                                                                          },
                                                                             "amount":  {
                                                                                            "type":  "number",
                                                                                            "format":  "double"
                                                                                        }
                                                                         }
                                                      },
...
                                       "Body":  {
                                                    "required":  [
                                                                     "name"
                                                                 ],
                                                    "type":  "object",
                                                    "properties":  {
                                                                       "name":  {
                                                                                    "type":  "string"
                                                                                }
                                                                   }
                                                },
...
                                       "Category":  {
                                                        "type":  "object",
                                                        "properties":  {
                                                                           "id":  {
                                                                                      "type":  "integer",
                                                                                      "format":  "int64"
                                                                                  },
                                                                           "name":  {
                                                                                        "type":  "string"
                                                                                    }
                                                                       }
                                                    },
...
                                   }
                   }
}